+++ packages/common/src/AcCmObject.ts
@@ -1,6 +1,5 @@
-import { clone, defaults, has, isEmpty, isEqual } from 'lodash-es'
-
 import { AcCmEventManager } from './AcCmEventManager'
+import { clone, defaults, has, isEmpty, isEqual } from './AcCmLodashUtils'
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type AcCmAttributes = Record<string, any>