+++ packages/graphic-interface/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mlightcad/graphic-interface
 
+## 1.0.3
+
+### Patch Changes
+
+- bundle common, geometry-engine, and graphic-interface into data-model and remove dependency on lodash-es
+- Updated dependencies
+  - @mlightcad/geometry-engine@1.0.3
+
 ## 1.0.2
 
 ### Patch Changes